add common.rs
